feat(step 13): Ajout des permissions avec bouncer

+++ api-suivi-eleves-etml/app/controllers/comments_controller.ts
@@ -2,6 +2,7 @@ import Comment from '#models/comment'
 import Student from '#models/student'
 import { commentValidator } from '#validators/comment'
 import type { HttpContext } from '@adonisjs/core/http'
+import CommentPolicy from '#policies/comment_policy'
 
 export default class CommentsController {
   /**
@@ -55,13 +56,20 @@ export default class CommentsController {
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ params, request, response, bouncer }: HttpContext) {
     const { content } = await request.validateUsing(commentValidator)
     // Vérifie que le commentaire appartient bien à l'élève
     const comment = await Comment.query()
       .where('id', params.id)
       .where('student_id', params.student_id)
       .firstOrFail()
+
+    if (await bouncer.with(CommentPolicy).denies('update', comment)) {
+      return response.unauthorized({
+        message:
+          "Vous n'êtes pas l'auteur de ce commentaire. Vous n'avez pas le droit de le modifier",
+      })
+    }
     // Mise à jour
     comment.content = content
     await comment.save()
@@ -72,11 +80,17 @@ export default class CommentsController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ params, response, bouncer }: HttpContext) {
     const comment = await Comment.query()
       .where('id', params.id)
       .where('student_id', params.student_id)
       .firstOrFail()
+    if (await bouncer.with(CommentPolicy).denies('delete', comment)) {
+      return response.unauthorized({
+        message:
+          "Vous n'êtes pas l'auteur de ce commentaire. Vous n'avez pas le droit de le supprimer",
+      })
+    }
     // Suppression du commentaire
     await comment.delete()
     // On utilise `response.noContent` pour retourner un code HTTP 204 sans contenu
